requirements.txt
Werkzeug==0.15.2
alembic==1.0.10
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
SQLAlchemy==1.2.12
marshmallow==2.15.6
ssdeep==3.3
psycopg2-binary==2.7.5
requests==2.21.0
apispec[yaml,validation]==1.3.3
bcrypt==3.1.4
pycodestyle==2.4.0
python-magic==0.4.15
luqum==0.7.4
logmatic-python==0.1.7
click==6.7
PyYAML==5.1.1
redis==3.2.1
